requirements.txt
python-telegram-bot==20.0
pandas==1.5.3
python-docx==0.8.11
python-dotenv==1.0.0
openpyxl==3.0.10
